MYSQL Database file
---- Problem Statement ---
Atliq Hardware is a leading manufacturer of computer hardware in India and other nations, but it struggles to make prompt and wise choices based on data. The management intends to add more junior data analysts to the data analytics team. Tony Sharma, the head of data analytics, conducts a SQL challenge to identify applicants with both technical and soft skills.